Overhaul of Client Side Static Pages to React.
I overhauled the entire front end in the client folder changing the static pages and translating the html into appropriate tsx file format for use as react components. Form here the Pages should be able to be rendered in a more single page application fashion.
Later on missing functionality from the transfer will be added in. The fox reviews page not loading with the jQuery functions even with importing jQuery so more research necessary there. If it proves too difficult the same functionality with css can be done through material UI or regular React.